♻️ refactor(cleanup): massive clean up and remove unwanted files and folders

+++ install.sh
@@ -75,13 +75,13 @@ install_wallpaper
 
 cleanup
 
-echo ":: Creating Hyprland desktop entry..."
+echo ":: Creating Hyprland desktop entry if doesn't exist..."
 
-curl https://raw.githubusercontent.com/hyprwm/Hyprland/refs/heads/main/systemd/hyprland-uwsm.desktop >hyprland-uwsm.desktop
-
-sudo cp hyprland-uwsm.desktop /usr/share/wayland-sessions/hyprland-uwsm.desktop
-
-rm hyprland-uwsm.desktop
+if [ ! -f /usr/share/wayland-sessions/hyprland-uwsm.desktop ]; then
+  curl https://raw.githubusercontent.com/hyprwm/Hyprland/refs/heads/main/systemd/hyprland-uwsm.desktop >hyprland-uwsm.desktop
+  sudo cp hyprland-uwsm.desktop /usr/share/wayland-sessions/hyprland-uwsm.desktop
+  rm hyprland-uwsm.desktop
+fi
 
 echo ":: Installation complete."
 echo "rebooting..."
